ChaoticEncryption.jl

Encrypt and decrypt image files using Pseudo-Random Number Generators!

Index

ChaoticEncryption.logistic_key

ChaoticEncryption.lorenz_key

ChaoticEncryption.substitution_decryption

ChaoticEncryption.substitution_encryption

Pseudo-Random Number Generators

ChaoticEncryption.logistic_key — Method

logistic_key(x_init, r, num_keys)

Generates pseudo-random keys using the Logistic Map.

The equation -

x_(n+1) = r * x_n * (1 - x_n)

Arguments

x_init::Float64: Initial value of x. x ϵ (0, 1).

r::Float64: A constant value. Values > 4 usually results in pseudo-random numbers.

num_keys::Int64: Number of keys to be generated.

Returns

keys::Array{Int64, 1}: Generated pseudo-random keys.

ChaoticEncryption.lorenz_key — Function

lorenz_key(x_init, y_init, z_init, num_keys, α=10.0, ρ=28.0, β=2.667, dt=0.01)

Generates 3 lists of pseudo-random numbers using Lorenz system of differential equations.

The equations -

dx/dt = α * (y - x)
dy/dt = x * (ρ - z) - y
dz/dt = x * y - β * z

Arguments

x_init::Float64: Initial value of x.

y_init::Float64: Initial value of y.

z_init::Float64: Initial value of z.

num_keys::Int64: Number of keys (in a single list) to be generated.

α::Float64: Constant associated with Lorenz system of differential equations.

ρ::Float64: Constant associated with Lorenz system of differential equations.

β::Float64: Constant associated with Lorenz system of differential equations.

Returns

x::Array{Int64, 1}: Generated pseudo-random keys corresponding to x values.

y::Array{Int64, 1}: Generated pseudo-random keys corresponding to y values.

z::Array{Int64, 1}: Generated pseudo-random keys corresponding to z values.

Encryption methods

ChaoticEncryption.substitution_encryption — Function

substitution_encryption(path_to_image, keys, path_for_result="./encrypted.png")

Performs substitution encryption on a given image with the given keys.

Algorithm

Iterates simulataneously over each pixel and key, and XORs the pixel value (all R, G, and B) with the given key. Hence, the order of the keys matter.

Arguments

path_to_image::String: The path to the image to be encrypted.

keys::Array{Int64, 1}: Keys for encryption.

path_for_result::String: The path for storing the encrypted image.

Example

julia> keys = logistic_key(0.01, 3.97, 719 * 718)
516242-element Array{Any,1}:
   0
  44
   7
  26
  14
 224
   ⋮
 115
  65
 126
 152
  74
 198
julia> substitution_encryption("D:\Saransh\Docs\PyCon_Squared.jpg", keys)
ENCRYPTING
ENCRYPTED

ChaoticEncryption.substitution_decryption — Function

substitution_decryption(path_to_image, keys, path_for_result="./decrypted.png")

Performs substitution decryption on a given image with the given keys.

Algorithm

Iterates simulataneously over each pixel and key, and XORs the pixel value (all R, G, and B) with the given key. Hence, the keys provided must be the same as the ones provided during encryption.

Arguments

path_to_image::String: The path to the image to be decrypted.

keys::Array{Int64, 1}: Keys for decryption.

path_for_result::String: The path for storing the decrypted image.

Example

julia> substitution_decryption("./encrypted.png", keys)
DECRYPTING
DECRYPTED