ChaoticEncryption.jl

Encrypt and decrypt image files using Pseudo-Random Number Generators!

Index

ChaoticEncryption.logistic_key

ChaoticEncryption.lorenz_key

ChaoticEncryption.substitution_decryption

ChaoticEncryption.substitution_encryption

Pseudo-Random Number Generators

ChaoticEncryption.logistic_key — Method

logistic_key(x_init, r, num_keys)

Generates pseudo-random keys using the Logistic Map.

The equation -

\[x_{n+1} = r * x_{n} * (1 - x_{n})\]

Arguments

x_init::Float64: Initial value of x. x ϵ (0, 1).

r::Float64: A constant value. Values > 4 usually results in pseudo-random numbers.

num_keys::Int64: Number of keys to be generated.

Returns

keys::Array{Int64, 1}: Generated pseudo-random keys.

Example

julia> using ChaoticEncryption

julia> logistic_key(0.01, 3.97, 20)
20-element Array{Int64,1}:
   0
  44
   7
  26
  14
 224
  16
 250
 162
 211
 200
 217
  97
 132
 134
 100
 135
 232
 122
 102

ChaoticEncryption.lorenz_key — Method

lorenz_key(x_init, y_init, z_init, num_keys; α=10.0, ρ=28.0, β=2.667, dt=0.01)

Generates 3 lists of pseudo-random numbers using Lorenz system of differential equations.

The equations -

\[\frac{dx}{dt} = α * (y - x)\]

\[\frac{dy}{dt} = x * (ρ - z) - y\]

\[\frac{dz}{dt} = x * y - β * z\]

Arguments

x_init::Float64: Initial value of x.

y_init::Float64: Initial value of y.

z_init::Float64: Initial value of z.

num_keys::Int64: Number of keys (in a single list) to be generated.

α::Float64: Constant associated with Lorenz system of differential equations.

ρ::Float64: Constant associated with Lorenz system of differential equations.

β::Float64: Constant associated with Lorenz system of differential equations.

Returns

x::Array{Int64, 1}: Generated pseudo-random keys corresponding to x values.

y::Array{Int64, 1}: Generated pseudo-random keys corresponding to y values.

z::Array{Int64, 1}: Generated pseudo-random keys corresponding to z values.

Example

julia> using ChaoticEncryption

julia> lorenz_key(0.01, 0.02, 0.03, 20)
([0, 0, 256, 24, 129, 42, 54, 134, 43, 179, 85, 19, 24, 44, 71, 210, 238, 152, 22, 27], [0, 0, 240, 55, 25, 163, 89, 243, 123, 5, 197, 64, 227, 54, 188, 226, 154, 134, 64, 69], [0, 0, 80, 227, 178, 204, 89, 33, 144, 139, 105, 208, 108, 155, 61, 254, 57, 102, 149, 47])

Encryption methods

ChaoticEncryption.substitution_encryption — Method

substitution_encryption(image, keys; path_for_result="./encrypted.png")

Performs substitution encryption on a given image with the given keys.

Algorithm

Iterates simulataneously over each pixel and key, and XORs the pixel value (all R, G, and B) with the given key. Hence, the order of the keys matter.

Arguments

image::Array{RGB{N0f8},2}: A loaded image.

keys::Array{Int64, 1}: Keys for encryption.

path_for_result::String: The path for storing the encrypted image.

Returns

encryptedImage::Array{RGB{N0f8}, 2}: Encrypted image.

Example

julia> using TestImages, ChaoticEncryption

julia> img = testimage("mandrill")
512×512 Array{RGB{N0f8},2} with eltype ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}:
 RGB{N0f8}(0.643,0.588,0.278)  …  RGB{N0f8}(0.702,0.737,0.463)
 RGB{N0f8}(0.471,0.49,0.243)      RGB{N0f8}(0.471,0.541,0.29)
 RGB{N0f8}(0.388,0.29,0.122)      RGB{N0f8}(0.376,0.314,0.192)
 RGB{N0f8}(0.251,0.31,0.188)      RGB{N0f8}(0.243,0.306,0.184)
 RGB{N0f8}(0.431,0.329,0.153)     RGB{N0f8}(0.314,0.369,0.235)
 RGB{N0f8}(0.38,0.329,0.114)   …  RGB{N0f8}(0.659,0.788,0.463)
 RGB{N0f8}(0.239,0.149,0.09)      RGB{N0f8}(0.651,0.675,0.467)
 RGB{N0f8}(0.212,0.09,0.122)      RGB{N0f8}(0.525,0.596,0.314)
 RGB{N0f8}(0.251,0.212,0.216)     RGB{N0f8}(0.467,0.518,0.29)
 RGB{N0f8}(0.216,0.125,0.09)      RGB{N0f8}(0.596,0.561,0.239)
 ⋮                             ⋱
 RGB{N0f8}(0.416,0.486,0.498)     RGB{N0f8}(0.267,0.365,0.345)
 RGB{N0f8}(0.459,0.51,0.463)      RGB{N0f8}(0.278,0.231,0.212)
 RGB{N0f8}(0.42,0.451,0.467)   …  RGB{N0f8}(0.329,0.337,0.365)
 RGB{N0f8}(0.38,0.42,0.447)       RGB{N0f8}(0.29,0.353,0.286)
 RGB{N0f8}(0.365,0.388,0.353)     RGB{N0f8}(0.349,0.325,0.333)
 RGB{N0f8}(0.376,0.451,0.525)     RGB{N0f8}(0.325,0.357,0.376)
 RGB{N0f8}(0.475,0.58,0.608)      RGB{N0f8}(0.318,0.314,0.235)
 RGB{N0f8}(0.494,0.663,0.659)  …  RGB{N0f8}(0.314,0.247,0.278)
 RGB{N0f8}(0.035,0.043,0.047)     RGB{N0f8}(0.016,0.02,0.008)

julia> height, width = size(img)
(512, 512)

julia> keys = logistic_key(0.01, 3.97, height * width)
262144-element Array{Int64,1}:
   0
  44
   7
  26
  14
 224
  16
 250
 162
 211
   ⋮
 185
 196
 214
  69
  22
  84
 234
 140
 253

julia> substitution_encryption(img, keys)
ENCRYPTING
ENCRYPTED
512×512 Array{RGB{N0f8},2} with eltype ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}:
 RGB{N0f8}(0.643,0.588,0.278)  …  RGB{N0f8}(0.969,0.973,0.196)
 RGB{N0f8}(0.016,0.004,0.259)     RGB{N0f8}(0.91,0.102,0.855)
 RGB{N0f8}(0.525,0.686,0.98)      RGB{N0f8}(0.447,0.259,0.137)
 RGB{N0f8}(0.043,0.016,0.482)     RGB{N0f8}(0.58,0.894,0.522)
 RGB{N0f8}(0.063,0.165,0.349)     RGB{N0f8}(0.78,0.788,0.671)
 RGB{N0f8}(0.475,0.298,0.02)   …  RGB{N0f8}(0.055,0.435,0.816)
 RGB{N0f8}(0.78,0.863,0.929)      RGB{N0f8}(0.243,0.204,0.937)
 RGB{N0f8}(0.392,0.271,0.302)     RGB{N0f8}(0.929,0.953,0.231)
 RGB{N0f8}(0.282,0.243,0.247)     RGB{N0f8}(0.718,0.267,0.541)
 RGB{N0f8}(0.396,0.447,0.271)     RGB{N0f8}(0.106,0.047,0.745)
 ⋮                             ⋱
 RGB{N0f8}(0.588,0.502,0.514)     RGB{N0f8}(0.992,0.894,0.882)
 RGB{N0f8}(0.29,0.741,0.286)      RGB{N0f8}(0.788,0.71,0.722)
 RGB{N0f8}(0.624,0.529,0.514)  …  RGB{N0f8}(0.855,0.847,0.827)
 RGB{N0f8}(0.467,0.49,0.392)      RGB{N0f8}(0.102,0.039,0.098)
 RGB{N0f8}(0.012,0.239,0.016)     RGB{N0f8}(0.176,0.153,0.129)
 RGB{N0f8}(0.929,0.996,0.043)     RGB{N0f8}(0.267,0.298,0.467)
 RGB{N0f8}(0.71,0.345,0.341)      RGB{N0f8}(0.204,0.208,0.349)
 RGB{N0f8}(0.902,0.192,0.188)  …  RGB{N0f8}(0.157,0.278,0.247)
 RGB{N0f8}(0.718,0.71,0.698)      RGB{N0f8}(0.976,0.973,1.0)

ChaoticEncryption.substitution_decryption — Method

substitution_decryption(image, keys; path_for_result="./decrypted.png")

Performs substitution decryption on a given image with the given keys.

Algorithm

Iterates simulataneously over each pixel and key, and XORs the pixel value (all R, G, and B) with the given key. Hence, the keys provided must be the same as the ones provided during encryption.

Arguments

image: ::String or ::Array{RGB{N0f8},2}. The path to the image or the loaded image to be decrypted.

keys::Array{Int64, 1}: Keys for decryption.

path_for_result::String: The path for storing the decrypted image.

Returns

decryptedImage::Array{RGB{N0f8}, 2}: Decrypted image.

Example

julia> using TestImages, ChaoticEncryption

julia> img = testimage("mandrill")
512×512 Array{RGB{N0f8},2} with eltype ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}:
 RGB{N0f8}(0.643,0.588,0.278)  …  RGB{N0f8}(0.702,0.737,0.463)
 RGB{N0f8}(0.471,0.49,0.243)      RGB{N0f8}(0.471,0.541,0.29)
 RGB{N0f8}(0.388,0.29,0.122)      RGB{N0f8}(0.376,0.314,0.192)
 RGB{N0f8}(0.251,0.31,0.188)      RGB{N0f8}(0.243,0.306,0.184)
 RGB{N0f8}(0.431,0.329,0.153)     RGB{N0f8}(0.314,0.369,0.235)
 RGB{N0f8}(0.38,0.329,0.114)   …  RGB{N0f8}(0.659,0.788,0.463)
 RGB{N0f8}(0.239,0.149,0.09)      RGB{N0f8}(0.651,0.675,0.467)
 RGB{N0f8}(0.212,0.09,0.122)      RGB{N0f8}(0.525,0.596,0.314)
 RGB{N0f8}(0.251,0.212,0.216)     RGB{N0f8}(0.467,0.518,0.29)
 RGB{N0f8}(0.216,0.125,0.09)      RGB{N0f8}(0.596,0.561,0.239)
 ⋮                             ⋱
 RGB{N0f8}(0.416,0.486,0.498)     RGB{N0f8}(0.267,0.365,0.345)
 RGB{N0f8}(0.459,0.51,0.463)      RGB{N0f8}(0.278,0.231,0.212)
 RGB{N0f8}(0.42,0.451,0.467)   …  RGB{N0f8}(0.329,0.337,0.365)
 RGB{N0f8}(0.38,0.42,0.447)       RGB{N0f8}(0.29,0.353,0.286)
 RGB{N0f8}(0.365,0.388,0.353)     RGB{N0f8}(0.349,0.325,0.333)
 RGB{N0f8}(0.376,0.451,0.525)     RGB{N0f8}(0.325,0.357,0.376)
 RGB{N0f8}(0.475,0.58,0.608)      RGB{N0f8}(0.318,0.314,0.235)
 RGB{N0f8}(0.494,0.663,0.659)  …  RGB{N0f8}(0.314,0.247,0.278)
 RGB{N0f8}(0.035,0.043,0.047)     RGB{N0f8}(0.016,0.02,0.008)

julia> height, width = size(img)
(512, 512)

julia> keys = logistic_key(0.01, 3.97, height * width)
262144-element Array{Int64,1}:
   0
  44
   7
  26
  14
 224
  16
 250
 162
 211
   ⋮
 185
 196
 214
  69
  22
  84
 234
 140
 253

julia> substitution_decryption("./encrypted.png", keys)
DECRYPTING
DECRYPTED
512×512 Array{RGB{N0f8},2} with eltype ColorTypes.RGB{FixedPointNumbers.Normed{UInt8,8}}:
 RGB{N0f8}(0.643,0.588,0.278)  …  RGB{N0f8}(0.702,0.737,0.463)
 RGB{N0f8}(0.471,0.49,0.243)      RGB{N0f8}(0.471,0.541,0.29)
 RGB{N0f8}(0.388,0.29,0.122)      RGB{N0f8}(0.376,0.314,0.192)
 RGB{N0f8}(0.251,0.31,0.188)      RGB{N0f8}(0.243,0.306,0.184)
 RGB{N0f8}(0.431,0.329,0.153)     RGB{N0f8}(0.314,0.369,0.235)
 RGB{N0f8}(0.38,0.329,0.114)   …  RGB{N0f8}(0.659,0.788,0.463)
 RGB{N0f8}(0.239,0.149,0.09)      RGB{N0f8}(0.651,0.675,0.467)
 RGB{N0f8}(0.212,0.09,0.122)      RGB{N0f8}(0.525,0.596,0.314)
 RGB{N0f8}(0.251,0.212,0.216)     RGB{N0f8}(0.467,0.518,0.29)
 RGB{N0f8}(0.216,0.125,0.09)      RGB{N0f8}(0.596,0.561,0.239)
 ⋮                             ⋱
 RGB{N0f8}(0.416,0.486,0.498)     RGB{N0f8}(0.267,0.365,0.345)
 RGB{N0f8}(0.459,0.51,0.463)      RGB{N0f8}(0.278,0.231,0.212)
 RGB{N0f8}(0.42,0.451,0.467)   …  RGB{N0f8}(0.329,0.337,0.365)
 RGB{N0f8}(0.38,0.42,0.447)       RGB{N0f8}(0.29,0.353,0.286)
 RGB{N0f8}(0.365,0.388,0.353)     RGB{N0f8}(0.349,0.325,0.333)
 RGB{N0f8}(0.376,0.451,0.525)     RGB{N0f8}(0.325,0.357,0.376)
 RGB{N0f8}(0.475,0.58,0.608)      RGB{N0f8}(0.318,0.314,0.235)
 RGB{N0f8}(0.494,0.663,0.659)  …  RGB{N0f8}(0.314,0.247,0.278)
 RGB{N0f8}(0.035,0.043,0.047)     RGB{N0f8}(0.016,0.02,0.008)